[sync] [Rust Frontend] Split engine core DTOs into separate modules (#47265)
Split engine-core protocol DTOs into request, sampling, output, structured_outputs, and dtype module paths; moved output decoding/classification ownership; updated Rust imports across the fork. Could not run cargo fmt/check because cargo/rustfmt are not installed in the environment.

Upstream-vLLM: c638f9216a08bfb5644d8a266ddd35421e04118d
