+++ rust/src/engine-core-client/examples/external_engine_logprobs.rs
@@ -5,9 +5,9 @@ use clap::Parser;
 use futures::StreamExt as _;
 use tokio::time::timeout;
 use tracing_subscriber::EnvFilter;
-use aphrodite_engine_core_client::protocol::{
-    EngineCoreFinishReason, EngineCoreRequest, EngineCoreSamplingParams,
-};
+use aphrodite_engine_core_client::protocol::output::EngineCoreFinishReason;
+use aphrodite_engine_core_client::protocol::request::EngineCoreRequest;
+use aphrodite_engine_core_client::protocol::sampling::EngineCoreSamplingParams;
 use aphrodite_engine_core_client::{
     EngineCoreClient, EngineCoreClientConfig, EngineCoreStreamOutput, TransportMode,
 };
