+++ rust/src/llm/src/request_metrics.rs
@@ -2,7 +2,7 @@ use std::sync::OnceLock;
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use aphrodite_engine_core_client::protocol::stats::PrefillStats;
-use aphrodite_engine_core_client::protocol::{EngineCoreEvent, EngineCoreEventType, EngineCoreOutput};
+use aphrodite_engine_core_client::protocol::output::{EngineCoreEvent, EngineCoreEventType, EngineCoreOutput};
 use aphrodite_metrics::{
     EngineLabels, FinishedReasonLabels, METRICS, PromptTokenSourceLabels, RequestMetrics,
 };
@@ -399,7 +399,7 @@ pub(crate) fn current_unix_timestamp_secs() -> f64 {
 #[cfg(test)]
 mod tests {
     use aphrodite_engine_core_client::protocol::stats::PrefillStats;
-    use aphrodite_engine_core_client::protocol::{EngineCoreEvent, EngineCoreEventType};
+    use aphrodite_engine_core_client::protocol::output::{EngineCoreEvent, EngineCoreEventType};
 
     use super::{RequestMetricsTracker, diff_or_zero};
 
@@ -411,7 +411,7 @@ mod tests {
             2,
             10.0,
             100.2,
-            &aphrodite_engine_core_client::protocol::EngineCoreOutput {
+            &aphrodite_engine_core_client::protocol::output::EngineCoreOutput {
                 request_id: "req-1".to_string(),
                 new_token_ids: vec![1],
                 finish_reason: None,
@@ -439,7 +439,7 @@ mod tests {
             2,
             11.5,
             100.4,
-            &aphrodite_engine_core_client::protocol::EngineCoreOutput {
+            &aphrodite_engine_core_client::protocol::output::EngineCoreOutput {
                 request_id: "req-1".to_string(),
                 new_token_ids: vec![2, 3],
                 finish_reason: None,
