+++ rust/src/mock-engine/src/io.rs
@@ -5,9 +5,8 @@ use tokio_util::sync::CancellationToken;
 use tracing::warn;
 use aphrodite_engine_core_client::mock_engine::MockEngineDataSockets;
 use aphrodite_engine_core_client::protocol::utility::EngineCoreUtilityRequest;
-use aphrodite_engine_core_client::protocol::{
-    EngineCoreRequest, EngineCoreRequestType, decode_msgpack, encode_msgpack,
-};
+use aphrodite_engine_core_client::protocol::{decode_msgpack, encode_msgpack};
+use aphrodite_engine_core_client::protocol::request::{EngineCoreRequest, EngineCoreRequestType};
 use zeromq::{DealerSocket, PushSocket, SocketRecv as _, SocketSend as _, ZmqMessage};
 
 use crate::engine::{EngineInput, EngineOutput};
